requirements.txt
beautifulsoup4
pillow
markdown
pygments
pymdown-extensions
mdx-wikilink-plus
pyyaml
